Added method to produce json reports
features: we don't update files once they are marked final,
and we only mark them final if dekamer.be lists them as final and they were processed successfully

This is a step towards publishing plenary json files in a separate repository.
Once that is in place we can move the document summarization stuff to a private repo.